move App to components dir
